backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
torch>=2.6.0
transformers>=4.35.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
esprima>=4.0.1
tqdm>=4.66.1
wandb>=0.16.0
python-multipart>=0.0.6
supabase>=2.0.0
scikit-learn>=1.3.0 